+++ logs.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-. ./config.sh
-
-docker logs $DOCKER_YOBI_NAME